+++ src/services/playlist/index.js
@@ -1,5 +1,17 @@
 const youtubeService = require("./youtube");
 const spotifyService = require("./spotify");
 
+exports.splitArray = (array, offset) => {
+  let start = 0;
+  let end = offset;
+  const result = [];
+  while (start < array.length) {
+    result.push(array.slice(start, end));
+    start = end;
+    end += offset;
+  }
+  return result;
+};
+
 exports.youtubeService = youtubeService;
 exports.spotifyService = spotifyService;
